Update user/role responses to return empty arrays
