+++ influx/users_test.go
@@ -97,12 +97,12 @@ func TestClient_Add(t *testing.T) {
 		u   *chronograf.User
 	}
 	tests := []struct {
-		name      string
-		args      args
-		status    int
-		want      *chronograf.User
-		wantQuery string
-		wantErr   bool
+		name        string
+		args        args
+		status      int
+		want        *chronograf.User
+		wantQueries []string
+		wantErr     bool
 	}{
 		{
 			name:   "Create User",
@@ -114,10 +114,22 @@ func TestClient_Add(t *testing.T) {
 					Passwd: "Dont Need Roads",
 				},
 			},
-			wantQuery: `CREATE USER "docbrown" WITH PASSWORD 'Dont Need Roads'`,
+			wantQueries: []string{
+				`CREATE USER "docbrown" WITH PASSWORD 'Dont Need Roads'`,
+				`SHOW USERS`,
+				`SHOW GRANTS FOR "docbrown"`,
+			},
 			want: &chronograf.User{
-				Name:   "docbrown",
-				Passwd: "Dont Need Roads",
+				Name: "docbrown",
+				Permissions: chronograf.Permissions{
+					chronograf.Permission{
+						Scope: chronograf.AllScope,
+						Allowed: chronograf.Allowances{
+							"WRITE",
+							"READ",
+						},
+					},
+				},
 			},
 		},
 		{
@@ -130,19 +142,19 @@ func TestClient_Add(t *testing.T) {
 					Passwd: "Dont Need Roads",
 				},
 			},
-			wantQuery: `CREATE USER "docbrown" WITH PASSWORD 'Dont Need Roads'`,
-			wantErr:   true,
+			wantQueries: []string{`CREATE USER "docbrown" WITH PASSWORD 'Dont Need Roads'`},
+			wantErr:     true,
 		},
 	}
 	for _, tt := range tests {
-		query := ""
+		queries := []string{}
 		ts := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
 			if path := r.URL.Path; path != "/query" {
 				t.Error("Expected the path to contain `/query` but was", path)
 			}
-			query = r.URL.Query().Get("q")
+			queries = append(queries, r.URL.Query().Get("q"))
 			rw.WriteHeader(tt.status)
-			rw.Write([]byte(`{"results":[{}]}`))
+			rw.Write([]byte(`{"results":[{"series":[{"columns":["user","admin"],"values":[["admin",true],["docbrown",true],["reader",false]]}]}]}`))
 		}))
 		u, _ := url.Parse(ts.URL)
 		c := &Client{
@@ -155,9 +167,16 @@ func TestClient_Add(t *testing.T) {
 			t.Errorf("%q. Client.Add() error = %v, wantErr %v", tt.name, err, tt.wantErr)
 			continue
 		}
-		if tt.wantQuery != query {
-			t.Errorf("%q. Client.Add() query = %v, want %v", tt.name, query, tt.wantQuery)
+		if len(tt.wantQueries) != len(queries) {
+			t.Errorf("%q. Client.Add() queries = %v, want %v", tt.name, queries, tt.wantQueries)
+			continue
 		}
+		for i := range tt.wantQueries {
+			if tt.wantQueries[i] != queries[i] {
+				t.Errorf("%q. Client.Add() query = %v, want %v", tt.name, queries[i], tt.wantQueries[i])
+			}
+		}
+
 		if !reflect.DeepEqual(got, tt.want) {
 			t.Errorf("%q. Client.Add() = %v, want %v", tt.name, got, tt.want)
 		}
